+++ uniffi_bindgen/src/bindings/kotlin/templates/CustomTypeTemplate.kt
@@ -39,26 +39,26 @@ public typealias {{ type_name }} = {{ concrete_type_name }}
 public object {{ ffi_converter_name }}: FfiConverter<{{ type_name }}, {{ ffi_type_name }}> {
     override fun lift(value: {{ ffi_type_name }}): {{ type_name }} {
         val builtinValue = {{ builtin|lift_fn }}(value)
-        return {{ config.into_custom.render("builtinValue") }}
+        return {{ config.lift("builtinValue") }}
     }
 
     override fun lower(value: {{ type_name }}): {{ ffi_type_name }} {
-        val builtinValue = {{ config.from_custom.render("value") }}
+        val builtinValue = {{ config.lower("value") }}
         return {{ builtin|lower_fn }}(builtinValue)
     }
 
     override fun read(buf: ByteBuffer): {{ type_name }} {
         val builtinValue = {{ builtin|read_fn }}(buf)
-        return {{ config.into_custom.render("builtinValue") }}
+        return {{ config.lift("builtinValue") }}
     }
 
     override fun allocationSize(value: {{ type_name }}): ULong {
-        val builtinValue = {{ config.from_custom.render("value") }}
+        val builtinValue = {{ config.lower("value") }}
         return {{ builtin|allocation_size_fn }}(builtinValue)
     }
 
     override fun write(value: {{ type_name }}, buf: ByteBuffer) {
-        val builtinValue = {{ config.from_custom.render("value") }}
+        val builtinValue = {{ config.lower("value") }}
         {{ builtin|write_fn }}(builtinValue, buf)
     }
 }